fix(projects-service): Prevent Project owner removing themselves

+++ services/apps/projects/src/member/member.controller.ts
@@ -1,6 +1,7 @@
+import {AuthUser, UserToken} from '@app/keycloak-auth';
 import {NotFound} from '@app/not-found';
-import {Body, Controller, Delete, Get, Param, Put} from '@nestjs/common';
-import {ApiOkResponse, ApiTags} from '@nestjs/swagger';
+import {Body, ConflictException, Controller, Delete, Get, Param, Put} from '@nestjs/common';
+import {ApiConflictResponse, ApiOkResponse, ApiTags} from '@nestjs/swagger';
 import {ProjectAuth} from '../project/project-auth.decorator';
 import {ProjectService} from '../project/project.service';
 import {MemberAuth} from './member-auth.decorator';
@@ -10,6 +11,7 @@ import {MemberService} from './member.service';
 
 const forbiddenResponse = 'Not member of project';
 const forbiddenProjectResponse = 'Not owner of project.';
+const removeOwnerResponse = 'Owner cannot remove themselves as member.';
 
 @Controller('projects/:project/members')
 @ApiTags('Members')
@@ -55,10 +57,15 @@ export class MemberController {
   @ProjectAuth({forbiddenResponse: forbiddenProjectResponse})
   @NotFound()
   @ApiOkResponse({type: Member})
+  @ApiConflictResponse({description: removeOwnerResponse})
   async remove(
     @Param('project') project: string,
     @Param('user') user: string,
+    @AuthUser() token: UserToken,
   ): Promise<Member | null> {
+    if (token.sub === user) {
+      throw new ConflictException(removeOwnerResponse);
+    }
     return this.memberService.remove(project, user);
   }
 }
